account server

+++ account-server/src/main/java/project/model/Account.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -21,9 +23,9 @@ public class Account {
 
     private String iban;
 
-    private Double balance;
+    private BigDecimal balance;
 
-    private String currency; // AZN, USD, EUR
+    private String currency;
 
     private Long userId;
 
@@ -32,7 +34,4 @@ public class Account {
 
     @Column(nullable = false)
     private boolean active = true;
-
 }
-
-
